+++ src/modules/Header/MyHeader.module.scss
@@ -2,32 +2,12 @@
   display: flex;
   align-items: center;
   zoom: 1.4;
+  gap: 50px;
 
   .menu {
     flex: 1;
     min-width: 0;
     justify-content: center;
     zoom: 1.4;
-  }
-
-  .balance {
-    display: flex;
-    flex-direction: row-reverse;
-    align-items: center;
-    padding: 10px;
-    width: 300px;
-    height: 50px;
-    background-image: linear-gradient(
-      to right,
-      rgba(0, 28, 36, 0.7),
-      rgba(119, 136, 141, 0.822)
-    );
-    font-family: "Marvel", sans-serif;
-    font-size: 40px;
-    text-align: center;
-    color: #8de6ffe8;
-
-    border: 3px solid aqua;
-    border-radius: 1rem;
   }
 }
