+++ src/modules/Header/components/ValuesOutput/ValuesOutput.module.scss
@@ -1,0 +1,20 @@
+.value {
+  display: flex;
+  flex-direction: row-reverse;
+  align-items: center;
+  padding: 10px;
+  width: 300px;
+  height: 50px;
+  background-image: linear-gradient(
+    to right,
+    rgba(0, 28, 36, 0.7),
+    rgba(119, 136, 141, 0.822)
+  );
+  font-family: "Marvel", sans-serif;
+  font-size: 40px;
+  text-align: center;
+  color: #8de6ffe8;
+
+  border: 3px solid aqua;
+  border-radius: 1rem;
+}
