+++ src/store/sliced/incrementToClick.slice.ts
